+++ src/main/java/fm/api/rest/individual/interfaces/IIndividualDao.java
@@ -13,5 +13,5 @@ public interface IIndividualDao {
 
   void updateIndividual(IndividualPresenter model);
 
-  Boolean isIndividualExisting(Long userId);
+  Boolean isIndividualExisting(Integer userId);
 }
